+++ src/C3.Domain/Identity/IEntityIdGenerator.cs
@@ -1,0 +1,7 @@
+namespace C3.Domain.Identity
+{
+    public interface IEntityIdGenerator
+    {
+        EntityId<TAggregate> Next<TAggregate>();
+    }
+}